Don't take ChannelPromise as argument in Channel API.
Motivation:

At the moment the outbound operations of ChannelHandler take a ChannelPromise as argument. This ChannelPromise needs to be carried forward to the next handler in the pipeline until it hits the transport. This is API choice has a few quirks which we should aim to remove:

 - There is a difference between if you add a ChannelFutureListener to the ChannelPromise or the ChannelFuture that is returned by the outbound method in terms of the ordering of execution of the listeners. Sometimes we add the listener to the promise while in reality we usually always want to add it to the future to ensure the listerns are executed in the "correct order".
- It is quite easy to "loose" a promise by forgetting to use the right method which also takes a promise
- Outbound operations of a ChannelHandler currently have a signature which allws these to throw an Exception. This is odd as we really should fail the promise / future to signal an exception.
- We need to explicit handle the VoidChannelPromise
- We have no idea what EventExecutor is used for the passed in ChannelPromise which may invalid our assumption of threading.

While changing the method signature of the outbound operations of the ChannelHandler is a good step forward we should also take care of just remove all the methods from ChannelOutboundInvoker (and its sub-types) that take a ChannelPromise and just always use the methods that return a ChannelFuture only.

Modifications:

- Remove throws Exception from the outbound operations of ChannelHandler
- Change the signature of the methods that took a ChannelPromise to not take one anymore and just return a ChannelFuture
- Remove all operations for ChannelOutboundInvoker that take a ChannelPromise.
- Adjust all code to cope with the API changes

Result:

Cleaner API which is easier to reason about and easier to use.